feat: recruitmentRequest api (update,list, delete)

+++ app/services/RecruitmentRequest.py
@@ -2,6 +2,7 @@ from app.extensions import db
 from app.model import RecruitmentRequest, RecruitmentProgress
 from app.extensions import ma
 from flask_jwt_extended import jwt_required, get_jwt_identity,current_user
+from flask import abort
 
 
 class RecruitmentRequestService:
@@ -10,61 +11,75 @@ class RecruitmentRequestService:
 
     def get_all_recruitmentRequests():
         recruitmentRequests = RecruitmentRequest.query.all()
+        #pagination
+        # page = request.args.get('page', 1, type=int)
+        # per_page = request.args.get('per_page', 10, type=int)
+        # recruitmentRequests = RecruitmentRequest.query.paginate(page, per_page, False)
+        # next_url = url_for('api.recruitment_request', page=recruitmentRequests.next_num) \
+        #     if recruitmentRequests.has_next else None
+        # prev_url = url_for('api.recruitment_request', page=recruitmentRequests.prev_num) \
+        #     if recruitmentRequests.has_prev else None
+
+
         return recruitmentRequests
 
     def get_recruitmentRequest_by_id( id):
         recruitmentRequest = RecruitmentRequest.query.get(id)
+        if recruitmentRequest is None:
+            abort(404,'Recruitment Request {id} not found')
         return recruitmentRequest
 
     def create_recruitmentRequest(payload):
         print(payload)
-        position = payload['position']
-        jobDescription = payload['jobDescription']
-        city = payload['city']
-        department = payload['department']
-        recruitmentType = payload['recruitmentType']
-        jobDuties = payload['jobDuties']
-        requiredQualifications = payload['requiredQualifications']
-        salaryAndBenefit = payload['salaryAndBenefit']
-        expectedStartDate = payload['expectedStartDate']
-        headCount = payload['headCount']
+        position = payload['Position']
+        jobDescription = payload['JobDescription']
+        city = payload['City']
+        department = payload['Department']
+        recruitmentType = payload['RecruitmentType']
+        jobDuties = payload['JobDuties']
+        requiredQualifications = payload['RequiredQualifications']
+        salaryAndBenefit = payload['SalaryAndBenefit']
+        expectedStartDate = payload['ExpectedStartDate']
+        headCount = payload['HeadCount']
         
         current_user = get_jwt_identity()
         requesterId = current_user['id']
         hrId = requesterId
-        status = payload['status']
+        status = payload['Status']
         recruitmentRequest = RecruitmentRequest(position, jobDescription, city, department, recruitmentType, jobDuties, requiredQualifications, salaryAndBenefit, expectedStartDate, headCount, requesterId, hrId, status)
-        print(recruitmentRequest)
         db.session.add(recruitmentRequest)
         db.session.commit()
         return recruitmentRequest
 
-    def add_recruitmentRequest(recruitmentRequest):
-        db.session.add(recruitmentRequest)
-        db.session.commit()
-        return recruitmentRequest
 
     def update_recruitmentRequest( id, recruitmentRequest):
         recruitmentRequest_to_update = RecruitmentRequest.query.get(id)
-        recruitmentRequest_to_update.position = recruitmentRequest.position
-        recruitmentRequest_to_update.jobDescription = recruitmentRequest.jobDescription
-        recruitmentRequest_to_update.city = recruitmentRequest.city
-        recruitmentRequest_to_update.department = recruitmentRequest.department
-        recruitmentRequest_to_update.recruitmentType = recruitmentRequest.recruitmentType
-        recruitmentRequest_to_update.jobDuties = recruitmentRequest.jobDuties
-        recruitmentRequest_to_update.requiredQualifications = recruitmentRequest.requiredQualifications
-        recruitmentRequest_to_update.salaryAndBenefit = recruitmentRequest.salaryAndBenefit
-        recruitmentRequest_to_update.expectedStartDate = recruitmentRequest.expectedStartDate
-        recruitmentRequest_to_update.headCount = recruitmentRequest.headCount
-        recruitmentRequest_to_update.requesterId = recruitmentRequest.requesterId
-        recruitmentRequest_to_update.hrId = recruitmentRequest.hrId
-        recruitmentRequest_to_update.status = recruitmentRequest.status
-        recruitmentRequest_to_update.recruitmentProgressId = recruitmentRequest.recruitmentProgressId
-        db.session.commit()
+        if recruitmentRequest_to_update is None:
+            abort(404,'Recruitment Request {id} not found')
+        recruitmentRequest_to_update.Position = recruitmentRequest['Position']
+        recruitmentRequest_to_update.JobDescription = recruitmentRequest['JobDescription']
+        recruitmentRequest_to_update.City = recruitmentRequest['City']
+        recruitmentRequest_to_update.Department = recruitmentRequest['Department']
+        recruitmentRequest_to_update.RecruitmentType = recruitmentRequest['RecruitmentType']
+        recruitmentRequest_to_update.JobDuties = recruitmentRequest['JobDuties']
+        recruitmentRequest_to_update.RequiredQualifications = recruitmentRequest['RequiredQualifications']
+        recruitmentRequest_to_update.SalaryAndBenefit = recruitmentRequest['SalaryAndBenefit']
+        recruitmentRequest_to_update.ExpectedStartDate = recruitmentRequest['ExpectedStartDate']
+        recruitmentRequest_to_update.HeadCount = recruitmentRequest['HeadCount']
+        recruitmentRequest_to_update.Status = recruitmentRequest['Status']
+        try:
+            db.session.commit()
+        except Exception as e:
+            abort(400,'An error occurred while updating the recruitment request {e}')
         return recruitmentRequest_to_update
 
-    def delete_recruitmentRequest(self, id):
+    def delete_recruitmentRequest( id):
         recruitmentRequest = RecruitmentRequest.query.get(id)
-        db.session.delete(recruitmentRequest)
-        db.session.commit()
+        if recruitmentRequest is None:
+            abort(404, 'Recruitment Request {id} not found')
+        try:
+            db.session.delete(recruitmentRequest)
+            db.session.commit()
+        except Exception as e:
+            abort(400,'An error occurred while deleting the recruitment request {e}')
         return recruitmentRequest
